+++ packages/mailer/src/index.ts
@@ -1,16 +1,13 @@
-import type { MailerConfig } from "./types";
+import type { FastifyMailer, MailerConfig } from "./types";
 /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
 import type { ApiConfig } from "@dzangolab/fastify-config";
-import type { Transporter } from "nodemailer";
-
-export interface FastifyMailerNamedInstance {
-  [namespace: string]: Transporter;
-}
-
-export type FastifyMailer = FastifyMailerNamedInstance & Transporter;
 
 declare module "fastify" {
   interface FastifyInstance {
+    mailer: FastifyMailer;
+  }
+
+  interface FastifyRequest {
     mailer: FastifyMailer;
   }
 }
@@ -21,6 +18,16 @@ declare module "@dzangolab/fastify-config" {
   }
 }
 
+declare module "mercurius" {
+  interface MercuriusContext {
+    mailer?: FastifyMailer;
+  }
+}
+
 export { default } from "./plugin";
 
-export type { MailerConfig } from "./types";
+export type {
+  FastifyMailer,
+  FastifyMailerNamedInstance,
+  MailerConfig,
+} from "./types";